fix: rename files to camelCase
